Adding crash reporting to agent0
