+++ lib/chainsync/chainsync/db/hyperdrive/db_schema.py
@@ -4,9 +4,10 @@ from datetime import datetime
 from decimal import Decimal
 from typing import Union
 
-from chainsync.base.db_schema import Base
 from sqlalchemy import BigInteger, Boolean, DateTime, ForeignKey, Integer, Numeric, String
 from sqlalchemy.orm import Mapped, mapped_column
+
+from ..base import Base
 
 # pylint: disable=invalid-name
